pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rava"
version = "0.1.0"
description = "Regulatory-Aligned Verification for AI Agents research framework"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "RAVA Team" }]
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.8.2",
  "PyYAML>=6.0.1",
  "tqdm>=4.66.4",
  "numpy>=1.26.4",
  "pandas>=2.2.2",
  "python-dotenv>=1.0.1",
  "langchain>=0.3.27",
  "langchain-openai>=0.3.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]
full = [
  "datasets>=2.20.0",
  "kaggle>=1.6.17",
  "transformers>=4.43.0",
  "torch>=2.3.0",
]

[project.scripts]
rava = "rava.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
